- split auth context and provider into different filles
- added api alias
