+++ frontend/src/App.jsx
@@ -1,5 +1,5 @@
 import { BrowserRouter } from 'react-router-dom';
-import { AuthProvider } from '@contexts/AuthContext';
+import AuthProvider from '@contexts/AuthProvider';
 import AppRoutes from '@routes/AppRoutes';
 
 function App() {
